requirements.txt
flask
flask-caching
flask-login
Flask-Session
flask-sqlalchemy
flash
PyJWT
pandas
numpy
jsondiff
scipy==1.11.2
git+https://github.com/cmaclell/py_rete.git@htn